Add test to check project read exception
